+++ symbolic-apps/src/ZkFold/Symbolic/Apps/KYC.hs
@@ -12,14 +12,14 @@ import           ZkFold.Base.Algebra.Basic.Class  (FromConstant (fromConstant))
 import           ZkFold.Base.Algebra.Basic.Field  (Zp)
 import           ZkFold.Base.Algebra.Basic.Number
 import           ZkFold.Base.Data.Vector          (Vector, head, tail, toVector)
-import           ZkFold.Symbolic.Class            (Symbolic (BaseField))
+import           ZkFold.Symbolic.Class            (Symbolic)
 import           ZkFold.Symbolic.Data.Bool        (Bool, not, (&&))
 import           ZkFold.Symbolic.Data.ByteString  (ByteString, Resize (resize), concat, toWords)
-import           ZkFold.Symbolic.Data.Combinators (Ceil, GetRegisterSize, Iso (..), KnownRegisterSize, KnownRegisters,
+import           ZkFold.Symbolic.Data.Combinators (Iso (..), KnownRegisterSize, KnownRegisters,
                                                    RegisterSize (..))
 import           ZkFold.Symbolic.Data.Eq          (Eq ((==)), elem)
 import           ZkFold.Symbolic.Data.Ord         (Ord ((>=)))
-import           ZkFold.Symbolic.Data.UInt        (OrdWord, UInt)
+import           ZkFold.Symbolic.Data.UInt        (UInt)
 import           ZkFold.Symbolic.Interpreter      (Interpreter)
 
 type KYCByteString context = ByteString 256 context
@@ -64,7 +64,6 @@ kycExample :: forall n r rsc context . (
   , Eq (KYCHash context)
   , KnownRegisterSize r
   , KnownRegisters context 64 r
-  , KnownNat (Ceil (GetRegisterSize (BaseField context) 64 r) OrdWord)
   ) => KYCData n context -> KYCHash context -> Bool context
 kycExample kycData hash =
   let
